#76 add Decimal handling into TradePair class

+++ data/TradePair.py
@@ -3,6 +3,7 @@ from utils.other_utils import get_next_id
 from enums.deal_type import get_order_type_by_id
 from utils.exchange_utils import get_fee_by_exchange
 from utils.string_utils import float_to_str
+from decimal import Decimal
 
 
 class TradePair(BaseData):
@@ -35,5 +36,5 @@ class TradePair(BaseData):
     def compute_profit(deal_1, deal_2):
         # FIXME Profit in BTC = vol*price_ex1*(1-fee1) - vol*price_ex2 - vol*fee2*avg_price
         # ( We can take avg_price as (price_ex1+price_ex2)/2)
-        return deal_1.volume * deal_1.price * 0.01 * (100 - get_fee_by_exchange(deal_1.exchange_id)) - \
-               deal_2.volume * deal_2.price * 0.01 * (100 + get_fee_by_exchange(deal_2.exchange_id))
+        return deal_1.volume * deal_1.price * Decimal(0.01 * (100 - get_fee_by_exchange(deal_1.exchange_id))) - \
+               deal_2.volume * deal_2.price * Decimal(0.01 * (100 + get_fee_by_exchange(deal_2.exchange_id)))
